solver-service/requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
librosa
numpy
